[Decomp][Player] Cohesion split: extract instance methods to PlayerInstance.cpp
